+++ cmd/redirect/helpers_test.go
@@ -1,7 +1,6 @@
-package helpers
+package main
 
 import (
-	"net/http"
 	"testing"
 	"time"
 
@@ -35,33 +34,6 @@ func TestGetCountryCode(t *testing.T) {
 	for _, test := range tests {
 		got, _ := GetCountryCode(test.IP)
 		want := test.CountryCode
-
-		if got != want {
-			t.Errorf("got %q, wanted %q", got, want)
-		}
-	}
-
-}
-
-func TestGetHtmlTitle(t *testing.T) {
-
-	type addTest struct {
-		url, title string
-	}
-
-	tests := []addTest{
-		{"https://github.com/berabulut", "berabulut (Hüseyin Bera Bulut) · GitHub"},
-		{"https://www.digitalocean.com/", "DigitalOcean – The developer cloud"},
-	}
-
-	for _, test := range tests {
-		res, err := http.Get(test.url)
-		if err != nil {
-			t.Errorf("%s", err)
-		}
-
-		got, _ := GetHtmlTitle(res.Body)
-		want := test.title
 
 		if got != want {
 			t.Errorf("got %q, wanted %q", got, want)
