+++ web/templates/redirect.tmpl
@@ -10,14 +10,6 @@
       href="https://fonts.googleapis.com/css?family=Roboto:200,300,400,500,600,700&display=swap"
       rel="stylesheet"
     />
-    <link
-      href="https://fonts.googleapis.com/css?family=Open+Sans:200,300,400,500,600,700&display=swap"
-      rel="stylesheet"
-    />
-        <link
-      href="https://fonts.googleapis.com/css?family=Quicksand:200,300,400,500,600,700&display=swap"
-      rel="stylesheet"
-    />
     <title>Kapsule - Redirecting</title>
     <style>
       h1 {
